requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1

wheel==0.31.1

# Bleeding edge Django
django==2.1.1

# Configuration
django-environ==0.4.5

# Images
Pillow==3.2.0

# Python-PostgreSQL Database Adapter
psycopg2==2.7.5

# Time zones support
pytz==2018.5


# Your custom requirements go here

#Django-REST-Framework
djangorestframework==3.8.2
markdown==2.6.11
django-filter==2.0.0
djoser==1.3.0

django-cors-headers==2.4.0



pyopenssl==18.0.0
ndg-httpsclient==0.5.1
pyasn1==0.4.4

django-querycount==0.7.0

numpy==1.15.1
